requirements.txt
alembic
arrow
black
cachelib
Flask
Flask-Assets
Flask-APScheduler
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
Flake8
Flake8-pyproject
furl
pyScss
playwright
pydantic
pyhumps
pyyaml
requests
ruamel.yaml
sqlalchemy
sssimp
Wand
